requirements.txt
dash
dash-bio
dash-bootstrap-components
pandas
plotly
matplotlib
dash-cytoscape
gunicorn
openpyxl